+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Dockerfile updated to use Python 3.14.3-slim and no-cache pip install
 - Docker build test added to build script
 
+### Fixed
+
+- Updated and renamed several environment variables in documentation, and added unit test to verify documentation coverage for constants. (#151)
+
 ## [2026.3.21b2] - 2026-03-21
 
 ### Fixed